WIP, traceIds are always 0
